// src/interface/weeklyreport/type.ts
export interface IListLeave {
  id: number;
  created_date: string;
  from_date: string;
  holiday_status_id: HolidayStatusId;
  state: string;
  employee_id: EmployeeId;
  date_to: string;
  config_id: ConfigId;
  for_reasons: string;
  reasons: string;
  time: number;
  minutes: number;
}

export interface ConfigId {
  id: number | boolean;
  name: string | boolean;
}

export interface HolidayStatusId {
  id: number;
  name: string;
  type: string;
}

export interface EmployeeId {
  id: number;
  name: string;
}